Refactored Factorials to BigInteger

+++ test/edu/cnm/deepdive/FactorialsTest.java
@@ -2,6 +2,7 @@ package edu.cnm.deepdive;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+import java.math.BigInteger;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.function.Executable;
 
@@ -19,8 +20,8 @@ class FactorialsTest {
   void computeRecursive() {
     for (long[] testCase : testCases) {
       int n = (int) testCase[0];
-      long expected = testCase[1];
-      long actual = Factorials.computeRecursive(n);
+      BigInteger expected = BigInteger.valueOf(testCase[1]);
+      BigInteger actual = Factorials.computeRecursive(n);
       assertEquals(expected, actual);
     }
     assertThrows(IllegalArgumentException.class, () -> Factorials.computeRecursive(-1));
